+++ codegen/main.ts
@@ -1,24 +1,17 @@
-import { AnyObject, mapValues } from '@cbhq/cds-utils';
+import { mapValues } from '@cbhq/cds-utils';
 
 import { scaleConfig } from './configs/scaleConfig';
-import { Icon } from './icons/Icon';
 import { Palette } from './Palette';
 import { Spacing } from './Spacing';
 import { Spectrum } from './Spectrum/Spectrum';
 import { Type } from './Type/Type';
 import { TypeScript } from './Typescript';
-import { generateFromTemplate } from './utils/generateFromTemplate';
-import { logError } from './utils/logError';
+import { buildTemplates } from './utils/buildTemplates';
 import { docgen } from './website/docgen';
 import { updateTextStylesTable } from './website/updateTextStylesTable';
 
-async function loadTemplates(): Promise<
-  Record<string, { dest: string; data: Record<string, unknown>; config?: AnyObject }[]>
-> {
-  const iconData = await Icon.data();
-  const webIconData = await Icon.web();
-
-  return {
+(async function () {
+  const templates = {
     'css.ejs': [
       {
         dest: 'web/Text/textStyles.ts',
@@ -82,66 +75,13 @@ async function loadTemplates(): Promise<
         dest: 'web/styles/palette.ts',
         data: { palette: Palette.cssVariables },
       },
-      {
-        dest: 'mobile/Icon/iconGlyphMap.ts',
-        data: { iconGlyphMap: iconData.glyphMap },
-      },
-      {
-        dest: 'website/docs/components/examples/Icon/data.ts',
-        data: { iconNames: webIconData.names, iconSizes: webIconData.iconSizes },
-      },
-      ...webIconData.svgPaths,
     ],
     'docgen.ejs': docgen,
-    'typescript.ejs': [
-      ...TypeScript,
-      {
-        dest: 'common/types/IconSize.ts',
-        data: {
-          types: {
-            IconSize: iconData.sizes,
-            IconPixelSize: iconData.pixels,
-          },
-        },
-      },
-      {
-        dest: 'common/types/IconName.ts',
-        data: {
-          types: {
-            IconName: iconData.names,
-          },
-        },
-      },
-    ],
+    'typescript.ejs': TypeScript,
   };
-}
 
-(async function () {
-  const templateInputs: {
-    template: string;
-    dest: string;
-    data: Record<string, unknown>;
-    config?: AnyObject;
-  }[] = [];
+  Palette.validate();
+  await updateTextStylesTable();
 
-  try {
-    Palette.validate();
-
-    await updateTextStylesTable();
-
-    Object.entries(await loadTemplates()).forEach(([template, items]) => {
-      items.map(({ dest, data, config = {} }) => {
-        templateInputs.push({
-          template,
-          dest,
-          data,
-          config,
-        });
-      });
-    });
-
-    await Promise.all(templateInputs.map(generateFromTemplate));
-  } catch (err) {
-    logError(err);
-  }
+  await buildTemplates(templates);
 })();