+++ mobile/icons/index.ts
@@ -1,0 +1,3 @@
+export * from './Icon';
+export * from './FiatIcon';
+export * from './iconGlyphMap';